+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - A bug which stopped ICO files from working as custom hot images in the buttons
   toolbar was fixed. [[#547](https://github.com/reupen/columns_ui/pull/547)]
 
+- A GDI object leak when using ICO files as custom button images was fixed.
+  [[#547](https://github.com/reupen/columns_ui/pull/547)]
+
 ## 2.0.0-alpha.1
 
 ### Features